visualization code of declination slices
visualization1 uses approximations to plot the declination slice and superposes the voids injected in the mocks with the ones detected by VoidFInder in addition to the galaxies of the catalog. visualization2 does the same job using a more rigorous analytical method in computing the slices of voids that intersect the declination desired. For questions or suggestions, e-mail fzaidoun@u.rochester.edu